+++ src/parsers/then/literal.rs
@@ -4,7 +4,10 @@ use nom::error::{ErrorKind, FromExternalError};
 use nom::IResult;
 
 use super::ThenWrapper;
-use crate::{BuildExecute, BuildPropagate, CommandArgument, CommandError, Execute, Propagate, TaskLogic, TaskLogicNoArgs, Then};
+use crate::{
+    prefix, BuildExecute, BuildPropagate, Chain, ChildUsage, CommandArgument, CommandError, Execute, IntoMultipleUsage, MultipleUsage,
+    Prefix, Propagate, TaskLogic, TaskLogicNoArgs, Then,
+};
 
 /// Default [`Then`] implementation for argument parsers that return `()`.
 pub struct LiteralThen<A, E> {
@@ -41,6 +44,25 @@ where
             task,
         }
     }
+}
+
+impl<A, E> IntoMultipleUsage for LiteralThen<A, E>
+where
+    A: IntoMultipleUsage + ChildUsage,
+    E: IntoMultipleUsage,
+{
+    type Item = Prefix<(A::Child, &'static str), E::Item>;
+
+    fn usage_gen(&self) -> Self::Item { prefix((self.argument.usage_child(), " "), self.executor.usage_gen()) }
+}
+
+impl<A, E> ChildUsage for LiteralThen<A, E>
+where
+    A: ChildUsage,
+{
+    type Child = A::Child;
+
+    fn usage_child(&self) -> Self::Child { self.argument.usage_child() }
 }
 
 impl<A, E, U> Execute<U> for LiteralThen<A, E>
@@ -137,4 +159,35 @@ where
             },
         ))(input)
     }
+}
+
+impl<A, E, C> CommandArgument<()> for LiteralThenExecutor<A, E, C>
+where
+    A: CommandArgument<()>,
+{
+    fn parse<'a>(&self, input: &'a str) -> IResult<&'a str, (), CommandError<'a>> { self.argument.parse(input) }
+}
+
+impl<A, E, C> IntoMultipleUsage for LiteralThenExecutor<A, E, C>
+where
+    A: IntoMultipleUsage + ChildUsage,
+    E: IntoMultipleUsage,
+{
+    type Item = Chain<A::Item, Prefix<(A::Child, &'static str), E::Item>>;
+
+    fn usage_gen(&self) -> Self::Item {
+        self.argument
+            .argument
+            .usage_gen()
+            .chain(prefix((self.argument.argument.usage_child(), " "), self.argument.executor.usage_gen()))
+    }
+}
+
+impl<A, E, C> ChildUsage for LiteralThenExecutor<A, E, C>
+where
+    LiteralThen<A, E>: ChildUsage,
+{
+    type Child = <LiteralThen<A, E> as ChildUsage>::Child;
+
+    fn usage_child(&self) -> Self::Child { self.argument.usage_child() }
 }